==> analysis/1.1. Dataset 1 - Preparation.ipynb <==
# Dataset 1 - Preparation
---

from pathlib import Path
path = Path().absolute()

## Load original dataset
---

import re

data = []

with open(path / "assets/datasets/original-1.txt") as file:
    for line in file.read().splitlines():
        # remove suffix ";"
        # split data by ";"
        data.append(
            re.sub(";$",'',line)
            .split(';')
        )

## Create dataframe
---

import pandas as pd

df = pd.DataFrame(data, columns=["texts","targets"])

df.info()

df.fillna('', inplace=True)

df["targets"].value_counts()

### Normalized classes
---

true_labels = {
    "cinta": [
        "cinta kesayangannya",
        "sayang"
    ],
    "gembira": [
        "kegembiraanjo",
        "kegembiraan jo",
        "kegembiraan now",
        "kegembiraan saat ini",
        "kegembiraan saya",
        "kegembiraan them",
        "kegembiraan upcoming",
        "kegembiraan world",
        "kegembiraan yang glamor",
        "kegembiraan",
        "kegembiraan",
        "kegembiraanjo",
        "senang",
        "kegembiraan",
        "kegembiraanjo"
    ],
    "kaget": [
        "kejutan",
        "kejutansur",
        "kejutanyang luar biasa dan menyentuh",
        "terkejut"
    ],
    "marah": [
        "amarahmu",
        "kemarahan",
        "kemarahan",
        "kemarahanan",
        "kemarahanan"
    ],
    "sedih": [
        "kesedihan",
        "kesedihan",
        "kesedihanad",
        "kesedihanini.",
        "kesedihanlonely",
        "kesedihanlove",
        "kesedihans",
        "kesedihanyang diperlukan",
        "kesedihan"
    ],
    "sukacita": [
        "bersukacita",
        "sukacitajo"
    ],
    "takut": [
        "ketakutan intricate",
        "ketakutan",
        "ketakutanfe"
    ]
}

for i in range(len(df)):
    target: str = df.at[i,"targets"]
    target = target.replace(u'\u200b','')
    target = target.strip()

    text: str = df.at[i,"texts"]
    text = text.replace(u'\u200b','')
    text = text.strip()

    for true_label, labels in true_labels.items():
        if target in labels:
            target = true_label

    df.at[i,"targets"] = target
    df.at[i,"texts"] = text

df["targets"].value_counts()

### Delete rows with defective and low frequency target emotion
---

df.drop(df[df["targets"].isin(['',"mengherankan"])].index, inplace=True)

df["targets"].value_counts()

## Export normalize dataset
---

df.info()

df.to_csv(path / "assets/datasets/dataset-1.csv", sep=";", index=False)

## Split dataset into training set and testing set
---

X = list(df["texts"])
y = list(df["targets"])

from sklearnex.model_selection import train_test_split

X_train, X_test, y_train, y_test = train_test_split(
    X,
    y,
    test_size=0.2,
    random_state=42,
    stratify=y
)

pd.DataFrame([X_train, y_train]).transpose().set_axis(["texts","targets"], axis="columns").to_csv(path / "assets/datasets/training-set-1.csv", sep=";", index=False)
pd.DataFrame([X_test, y_test]).transpose().set_axis(["texts","targets"], axis="columns").to_csv(path / "assets/datasets/testing-set-1.csv", sep=";", index=False)
