==> analysis/2.1. Dataset 1 - Exploratory Data Analysis .ipynb <==
# Exploratory Data Analysis - Dataset 1
---

import pickle
from tqdm import tqdm

from pathlib import Path

path = Path().absolute()

## Load stanza model
---

# import stanza

# stanza.download("id")

# tokenizer = stanza.Pipeline("id", processors="tokenize,mwt,pos,lemma", use_gpu=True)

### Load dataset
---

import pandas as pd

df = pd.read_csv(path / "assets/datasets/dataset-1.csv", delimiter=";")

# X = df["texts"]

## Process Data
---

### Tokenization
---

# docs = tokenizer.bulk_process(X)

### Reduce docs
---

# for i, doc in enumerate(docs):
#     tokens = []

#     for sentence in doc.sentences:
#         for token in sentence.tokens:
#             tokens.extend(token.words)

#     docs[i] = (doc.text, tokens)

### Save tokenized docs
---

# with open(path / "assets/pickles/docs-1.pickle", "wb") as f:
#     pickle.dump(docs, f)

## Explore Data
---

### Load tokenized docs
---

with open(path / "assets/pickles/docs-1.pickle", "rb") as f:
    docs = pickle.load(f)

### Foreign word identification
---

# unique_words = set()

# for text, tokens in tqdm(docs):
#     for token in tokens:
#         unique_words.add(token.lemma)

# from kbbi import KBBI, TidakDitemukan

# foreign_words = set()

# for word in tqdm(unique_words):
#     try:
#         KBBI(word)
#     except TidakDitemukan:
#         foreign_words.add(word)

### Most common words
---

# lemmas = []

# for text, tokens in tqdm(docs):
#     for token in tokens:
#         lemmas.append(token.lemma)

# from collections import Counter

# Counter(lemmas).most_common()

### Part of Speech List
---

list_of_pos = set()

for text, tokens in docs:
    for token in tokens:
        if token.pos not in list_of_pos:
            list_of_pos.add(token.pos)

list_of_pos

### Most common tokens of specific POS tag
---

from collections import Counter

words=[]

for text, tokens in docs:
    for token in tokens:
        if token.pos == "X":
            words.append(token.text)

counter=Counter(words).most_common()

print(len(counter))
counter

words_2 = []

for text, tokens in docs:
    for token in tokens:
        if token.text == "love":
            print(token)
