==> analysis/4.2. Analysis 2.ipynb <==
---

from pathlib import Path

path = Path().absolute()

cd ../

import pickle
import time

---

### Instantiate Classifier
---

from pipeline.classification import Classification

clf = Classification()

### Load Data
---

# import pandas as pd

# # Load dataset
# training_set_df = pd.read_csv(path / "assets/datasets/training-set-1.csv", delimiter=";")

# # Get X and y from dataset
# X_train = list(training_set_df["texts"])
# y_train = list(training_set_df["targets"])

#### Get random samples
---

# from random import sample

# sample([x for x in X_train if len(x) <= 50 and "gugup" in x], 5)

#### Example of a dataset
---

X_temp = [
    "<b>Saya senang bisa melihat 🎆 bersamamu</b>",
    "Saya merasa benar-benar dirugikan dirinya!!!!",
    "Video menyeramkan lainnya: https://www.youtube.com/shorts/ZIl2iiHy1oM. ***JANGAN DILIHAT KALAU TIDAK BERANI***",
    "Saya merasa kesepian dan terisolasi akhir akhir ini",
    "saya merasa sangat sedih saya berharap saya mati",
    "saya merasa sedih hanya memikirkannya",
    "saya merasa patah hati dan sedih",
    "saya merasa sangat putus asa dan sedih",
    "Saya selalu merasa tidak diinginkan dan sedih",
    "saya merasa senang",
    "saya merasa cukup senang minggu ini",
    "saya sangat senang tapi saya merasa lelah",
    "saya merasakan cinta dan perhatian sekarang",
    "saya merasa disambut dan dicintai",
    "saya merasa dicintai",
    "saya merasa takut berada di dekat mereka",
    "saya merasa sangat ketakutan dan gugup",
    "saya merasa takut dan bodoh",
    "saya merasa takut",
    "saya sangat suka perasaan takut",
    "saya merasa seperti orang bodoh",
    "saya merasa bodoh tapi senang",
    "saya merasa sangat bodoh karena jujur",
    "saya merasa sangat bodoh dan malu",
    "saya merasa bodoh dan tidak diinginkan",
    "saya merasa kesal dan marah",
    "saya berteriak dengan perasaan sangat marah",
    "saya bahkan tidak bisa merasa marah karenanya",
    "saya merasa sedikit kesal hari ini",
    "saya bangun hari ini merasa kesal",
    "saya merasa sangat gemetar dan sedih",
    "saya menjadi bodoh merasa gugup"
]

y_temp = [
    "gembira",
    "marah",
    "takut",
    "sedih",
    "sedih",
    "sedih",
    "sedih",
    "sedih",
    "sedih",
    "gembira",
    "gembira",
    "gembira",
    "cinta",
    "cinta",
    "cinta",
    "takut",
    "takut",
    "takut",
    "takut",
    "takut",
    "sedih",
    "gembira",
    "sedih",
    "sedih",
    "sedih",
    "marah",
    "marah",
    "marah",
    "marah",
    "marah",
    "sedih",
    "sedih"
]

#### Split dataset
---

from sklearn.model_selection import train_test_split

X_temp_train, X_temp_test, y_temp_train, y_temp_test = train_test_split(
    X_temp,
    y_temp,
    test_size=0.2,
    random_state=42,
    stratify=y_temp
)

X_temp_train, X_temp_validation, y_temp_train, y_temp_validation = train_test_split(
    X_temp_train,
    y_temp_train,
    test_size=0.21,
    random_state=42,
    stratify=y_temp_train
)

print(len(X_temp_train))

# for x in X_temp_train:
#     print(x)

for y in y_temp_train:
    print(y)

print(len(X_temp_validation))

# for x in X_temp_validation:
#     print(x)

for y in y_temp_validation:
    print(y)

print(len(X_temp_test))

# for x in X_temp_test:
#     print(x)

for y in y_temp_test:
    print(y)

#### Text Preprocessing
---

X_temp_train_docs = clf.text_preprocessing_pipeline.transform(X_temp_train)
X_temp_validation_docs = clf.text_preprocessing_pipeline.transform(X_temp_validation)
X_temp_test_docs = clf.text_preprocessing_pipeline.transform(X_temp_test)

#### Feature Selection
---

clf.feature_selection_pipeline.named_steps["pos_filter"].set_params(**{"pos": set(["ADJ","ADV","NOUN","PART","VERB"])})
clf.feature_selection_pipeline.named_steps["document_transformer"].set_params(**{"feat_attrs": ["lemma","upos"]})
X_temp_train_selected = clf.feature_selection_pipeline.transform(X_temp_train_docs)
X_temp_validation_selected = clf.feature_selection_pipeline.transform(X_temp_validation_docs)
X_temp_test_selected = clf.feature_selection_pipeline.transform(X_temp_test_docs)

for i, doc in enumerate(X_temp_train_selected):
    print(", ".join(doc))

for i, doc in enumerate(X_temp_validation_selected):
    print(", ".join(doc))

for i, doc in enumerate(X_temp_test_selected):
    print(", ".join(doc))

#### Feature Extraction
---

tfidfvectorizer_hyperparameters = {
    "ngram_range": (1, 1),
    "max_df": (1.0),
    "min_df": (1),
    "norm": None,
    "sublinear_tf": False
}
tfidfvectorizer = clf.classification_pipeline.named_steps["tfidfvectorizer"]
tfidfvectorizer.set_params(**tfidfvectorizer_hyperparameters)
tfidfvectorizer.fit(X_temp_train_selected)

for i, zip_ in enumerate(
    zip(
        dict(
            sorted(
                tfidfvectorizer.vocabulary_.items(),
                key=lambda x: x[1]
            )
        ).items(),
        tfidfvectorizer.idf_
    )
):
    item, idf = zip_
    k, v = item
    df = 0

    for doc in X_temp_train_selected:
        if k in doc:
            df+=1

    print(i+1, k, df, round(idf, 5), sep="\t")
    # print(i+1, k, df, sep="\t")

X_temp_train_features = tfidfvectorizer.transform(X_temp_train_selected)
X_temp_validation_features = tfidfvectorizer.transform(X_temp_validation_selected)
X_temp_test_features = tfidfvectorizer.transform(X_temp_test_selected)

#### Multi classification
---

linearsvc_hyperparameters = {
    "penalty": "l2",
    "loss": "squared_hinge",
    "dual": False,
    "tol": 0.0001,
    "C": 0.01,
    "multi_class": "ovr",
    "fit_intercept": True,
    "intercept_scaling": 1.0,
    "class_weight": "balanced",
    "max_iter": 1000
}
linearsvc = clf.classification_pipeline.named_steps["linearsvc"]
linearsvc.set_params(**linearsvc_hyperparameters)
linearsvc.fit(X_temp_train_features, y_temp_train)

y_temp_validation_pred = linearsvc.predict(X_temp_validation_features)
y_temp_test_pred = linearsvc.predict(X_temp_test_features)

#### Evaluation
---

from sklearn.metrics import accuracy_score, matthews_corrcoef

# test
print(accuracy_score(y_temp_validation, y_temp_validation_pred))
print(matthews_corrcoef(y_temp_validation, y_temp_validation_pred))

# test
print(accuracy_score(y_temp_test, y_temp_test_pred))
print(matthews_corrcoef(y_temp_test, y_temp_test_pred))

y_temp_test_pred

clf.classification_pipeline.named_steps['linearsvc'].classes_

coef=clf.classification_pipeline.named_steps['linearsvc'].coef_
coef

intercept=clf.classification_pipeline.named_steps['linearsvc'].intercept_
intercept

import numpy as np

i=18
np.dot(X_temp_train_features[i].toarray(), coef.T) + intercept.T

np.dot(X_temp_train_features.toarray(), coef[1])

coef[1].T.shape

X_temp_train_features.toarray().shape

intercept[1]

coef[1]

(np.dot(coef[1][1:], X_temp_train_features[-1].toarray()[0][1:]) - intercept[1]) / coef[1][0]

(np.dot([v for i, v in enumerate(coef[1]) if i != 1], [v for i, v in enumerate(X_temp_train_features[-1].toarray()[0]) if i != 1]) - intercept[1]) / coef[1][1]

(np.dot(coef[1][:-1], X_temp_train_features[-1].toarray()[0][:-1]) - intercept[1]) / coef[1][-1]

2 / np.dot(coef[1].T, coef[1])
