==> analysis/4.3. Analysis 3.ipynb <==
---

from pathlib import Path

path = Path().absolute()

cd ../

import pickle
import time

---

### Instantiate Classifier
---

from pipeline.classification import Classification
clf = Classification()

### Load Data
---

with open(path / "assets/stopwords/default.txt") as f:
    stopwords = [line.strip() for line in f.readlines()]

### Load model
---

filepath = path / "assets/models/model.2023.06.10.14.19.44.007537.pickle"
clf.from_disk(filepath)

### Data Preparation
---

# Text Preprocessing
stopwords_processed = clf.text_preprocessing_pipeline.transform(stopwords)

# # Feature Selection
stopwords_processed = clf.feature_selection_pipeline.transform(stopwords_processed)

stopwords_processed = [l for l in stopwords_processed if l]
y = []

for l in stopwords_processed:
    feature = clf.classification_pipeline.named_steps["tfidfvectorizer"].transform(l)
    y.append(clf.classification_pipeline.named_steps["linearsvc"].decision_function(feature))

for i, stopword in enumerate(stopwords_processed):
    if any(x > 0 for x in y[i].tolist()[0]):
        print(stopword, y[i])

vocab = clf.classification_pipeline.named_steps["tfidfvectorizer"].get_feature_names_out()
vocab = [x.lower() for x in vocab]

for s in stopwords:
    if s in vocab:
        print(s)
